requirements.txt
gin-config==0.5.0
imageio==2.25.1
imageio-ffmpeg==0.4.8
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
lpips==0.1.4
nerfacc==0.3.5
numpy==1.23.5
torchmetrics==0.11.1
tqdm==4.64.1
